Исправление - объем должен быть в float

+++ ProgProc1/Ball.cpp
@@ -9,8 +9,8 @@ void Out(ball& b, ofstream& ofst)
 {
 	ofst << "It is Ball: r = "	<< b.r;
 }
-int volume(ball& b)
+float volume(ball& b)
 {
-	int pi = 3.14;
+	float pi = 3.14;
 	return (4 / 3) * pi * b.r * b.r;
 }
